Financial System v0.1.0 FinancialSystem.Currency

Data structure to handle currency operations. All inputed currencies must be in compliance with ISO 4217.

Functions

new(alphabetic_code, numeric_code, decimal_places)

Creates a new currency with its respective alphabetic code, numeric code and decimal places.

If successfull, returns a tuple with :ok status and the created FinancialSystem.Currency. Otherwise, raises a RuntimeError when the currency is not valid

Examples

iex> FinancialSystem.Currency.new("BRL", 100, 2)
{:ok,
  %FinancialSystem.Currency{
    alphabetic_code: "BRL",
    decimal_places: 2,
    numeric_code: 100
  }}

valid?(currency)

Checks if the given currency is valid

Returns :ok if successfull. Raises a RuntimeError when the currency is not valid.

Examples

iex> cur1 = %FinancialSystem.Currency{alphabetic_code: "BRL", numeric_code: 100, decimal_places: 2}
iex> FinancialSystem.Currency.valid?(cur1)
:ok

valid?(alphabetic_code, numeric_code, decimal_places)

Checks if the alphabetic code, numeric code and decimal places are valid.

Returns :ok if successfull. Raises a RuntimeError when the currency is not valid.

Examples

iex> FinancialSystem.Currency.valid?("BRL", 100, 2)
:ok

valid_alphabetic_code?(alphabetic_code)

Checks if the alphabetic code is valid.

Returns a tuple with status :ok and the given code if successfull. Otherwise, returns a tuple with status :error and the error message.

Examples

iex> FinancialSystem.Currency.valid_alphabetic_code?("BRL")
{:ok, "BRL"}

valid_numeric_code?(numeric_code)

Checks if the numeric code is valid.

Returns a tuple with status :ok and the given code if successfull. Otherwise, returns a tuple with status :error and the error message.

Examples

iex> FinancialSystem.Currency.valid_numeric_code?(100)
{:ok, 100}