Financial System v0.1.0 FinancialSystem.Money

Data structure to handle money operations. It takes advantage of Elixir’s bignum arithmetics with the basic type integer to store the minor units of a money value given its currency.

Functions

add(money1, value)

Addition operation. Adds a given value in the string format to the given money.

Returns a new FinancialSystem.Money. Raises an error if different currencies.

Examples

iex> {:ok, original_currency} = FinancialSystem.Currency.new("BRL", 100, 2)
iex> {:ok, money} = FinancialSystem.Money.new("13.37", original_currency)
iex> FinancialSystem.Money.add(money, "6.16")
%FinancialSystem.Money{
  currency: %FinancialSystem.Currency{
    alphabetic_code: "BRL",
    decimal_places: 2,
    numeric_code: 100
  },
  minor_units: 1953,
  precision: 2
}

add_money(money1, money2)

Addition operation. Adds money2 to money1.

Returns a new FinancialSystem.Money. Raises an error if different currencies.

Examples

iex> {:ok, original_currency} = FinancialSystem.Currency.new("BRL", 100, 2)
iex> {:ok, money1} = FinancialSystem.Money.new("13.37", original_currency)
iex> {:ok, money2} = FinancialSystem.Money.new("6.16", original_currency)
iex> FinancialSystem.Money.add_money(money1, money2)
%FinancialSystem.Money{
  currency: %FinancialSystem.Currency{
    alphabetic_code: "BRL",
    decimal_places: 2,
    numeric_code: 100
  },
  minor_units: 1953,
  precision: 2
}

as_string(money)

Formats the money in a string.

Examples

iex> {:ok, original_currency} = FinancialSystem.Currency.new("BRL", 100, 2)
iex> {:ok, money} = FinancialSystem.Money.new("13.37", original_currency)
iex> FinancialSystem.Money.as_string(money)
"13.37"

multiply(money, value)

Multiplication operation. Multiply value of money by a given value in the string format.

Returns a new FinancialSystem.Money.

Examples

iex> {:ok, original_currency} = FinancialSystem.Currency.new("BRL", 100, 2)
iex> {:ok, money} = FinancialSystem.Money.new("13.37", original_currency)
iex> FinancialSystem.Money.multiply(money, "6.16")
%FinancialSystem.Money{
  currency: %FinancialSystem.Currency{
    alphabetic_code: "BRL",
    decimal_places: 2,
    numeric_code: 100
  },
  minor_units: 823592,
  precision: 4
}

new(amount, precision \\ 2, currency)

Creates a new money with the amount given in a string format, a default precision of 2 in its minor units and a given currency.

If successfull, returns a tuple with :ok status and the created FinancialSystem.Money. Otherwise, raises a RuntimeError when the amount is not valid

Examples

iex> {:ok, currency} = FinancialSystem.Currency.new("BRL", 100, 2)
iex> FinancialSystem.Money.new("13.37", currency)
{:ok,
  %FinancialSystem.Money{
    currency: %FinancialSystem.Currency{
      alphabetic_code: "BRL",
      decimal_places: 2,
      numeric_code: 100
    },
    minor_units: 1337,
    precision: 2
  }}

string_to_minor_units(amount, decimal_places)

Parse a given amount in the string format to the money’s structure compatible minor units. decimal_places is the desired precision

Returns the parsed minor units.

Examples

iex> FinancialSystem.Money.string_to_minor_units("13.37", 2)
1337

subtract(money1, value)

Subtraction operation. Subtracts a given value in the string format from the given money.

Returns a new FinancialSystem.Money. Raises an error if different currencies.

Examples

iex> {:ok, original_currency} = FinancialSystem.Currency.new("BRL", 100, 2)
iex> {:ok, money} = FinancialSystem.Money.new("13.37", original_currency)
iex> FinancialSystem.Money.subtract(money, "6.16")
%FinancialSystem.Money{
  currency: %FinancialSystem.Currency{
    alphabetic_code: "BRL",
    decimal_places: 2,
    numeric_code: 100
  },
  minor_units: 721,
  precision: 2
}

subtract_money(money1, money2)

Subtraction operation. Subtracts the money2 from the money1.

Returns a new FinancialSystem.Money. Raises an error if different currencies.

Examples

iex> {:ok, original_currency} = FinancialSystem.Currency.new("BRL", 100, 2)
iex> {:ok, money1} = FinancialSystem.Money.new("13.37", original_currency)
iex> {:ok, money2} = FinancialSystem.Money.new("6.16", original_currency)
iex> FinancialSystem.Money.subtract_money(money1, money2)
%FinancialSystem.Money{
  currency: %FinancialSystem.Currency{
    alphabetic_code: "BRL",
    decimal_places: 2,
    numeric_code: 100
  },
  minor_units: 721,
  precision: 2
}

valid_amount?(amount)

Check validity of given amount.

Returns true if valid, false otherwise

Examples

iex> FinancialSystem.Money.valid_amount?("13.37")
true