Financial System v0.1.0 FinancialSystem

Financial operations with arbitrary-precision.

FinancialSystem is able to convert money from one currency to another with one requirement: that the currency is in compliance with ISO 4217, that includes the currency’s alphabetic and numeric codes and the exponent that represents its decimal places.

Functions

convert(money, currency, exchange_rate)

Converts money from its original currency to the new currency given via the parameter currency, given the exchange_rate

Examples

iex> {:ok, original_currency} = FinancialSystem.Currency.new("BRL", 100, 2)
iex> {:ok, new_currency} = FinancialSystem.Currency.new("USD", 101, 2)
iex> {:ok, money} = FinancialSystem.Money.new("10.50", original_currency)
iex> FinancialSystem.convert(money, new_currency, "3.34")
%FinancialSystem.Money{
  currency: %FinancialSystem.Currency{
    alphabetic_code: "USD",
    decimal_places: 2,
    numeric_code: 101
  },
  minor_units: 350700,
  precision: 4
}

deposit(account, amount)

Deposits the given amount into the given account.

Returns a new account with the new balance.

Examples

iex> {:ok, currency} = FinancialSystem.Currency.new("BRL", 100, 2)
iex> {:ok, money} = FinancialSystem.Money.new("42.42", currency)
iex> {:ok, acc} = FinancialSystem.Account.new(123, money, "Arthur Dent")
iex> FinancialSystem.deposit(acc, "20.00")
{:ok,
%FinancialSystem.Account{
  balance: %FinancialSystem.Money{
    currency: %FinancialSystem.Currency{
      alphabetic_code: "BRL",
      decimal_places: 2,
      numeric_code: 100
    },
    minor_units: 6242,
    precision: 2
  },
  number: 123,
  owner: "Arthur Dent"
}}

transfer(from_account, to_account, amount)

Transfer the given amount from the from_account to the to_account if the first one has enough funds, i.e., a balance greater than the amount.

Returns both new accounts, each with their respectives new balances.

Examples

iex> {:ok, currency} = FinancialSystem.Currency.new("BRL", 100, 2)
iex> {:ok, money} = FinancialSystem.Money.new("42.42", currency)
iex> {:ok, money2} = FinancialSystem.Money.new("0.42", currency)
iex> {:ok, from_acc} = FinancialSystem.Account.new(123, money, "Arthur Dent")
iex> {:ok, to_acc} = FinancialSystem.Account.new(234, money2, "Ford Prefect")
iex> FinancialSystem.transfer(from_acc, to_acc, "20.00")
{:ok,
%FinancialSystem.Account{
  balance: %FinancialSystem.Money{
    currency: %FinancialSystem.Currency{
      alphabetic_code: "BRL",
      decimal_places: 2,
      numeric_code: 100
    },
    minor_units: 2242,
    precision: 2
  },
  number: 123,
  owner: "Arthur Dent"
},
%FinancialSystem.Account{
  balance: %FinancialSystem.Money{
    currency: %FinancialSystem.Currency{
      alphabetic_code: "BRL",
      decimal_places: 2,
      numeric_code: 100
    },
    minor_units: 2042,
    precision: 2
  },
  number: 234,
  owner: "Ford Prefect"
}}

transfer_split(from_account, accounts_with_percentages, amount)

Transfer the given amount from the from_account to the to_accounts if the first one has enough funds, i.e., a balance greater than the amount. to_accounts is a list of tuples with each tuple having the first value as FinancialSystem.Account and the second value as the respective percentage of the money that will be transfered to the corresponding account, from 0 to 1 in a string format.

Returns the new from_account and a list with the new to_accounts and each with their respectives new balances.

Examples

iex> {:ok, currency} = FinancialSystem.Currency.new("BRL", 100, 2)
iex> {:ok, money} = FinancialSystem.Money.new("42.42", currency)
iex> {:ok, money2} = FinancialSystem.Money.new("10.42", currency)
iex> {:ok, money3} = FinancialSystem.Money.new("15.42", currency)
iex> {:ok, from_acc} = FinancialSystem.Account.new(123, money, "Arthur Dent")
iex> {:ok, to_acc1} = FinancialSystem.Account.new(423, money2, "Ford Prefect")
iex> {:ok, to_acc2} = FinancialSystem.Account.new(423, money3, "Marvin")
iex> FinancialSystem.transfer_split(from_acc, [{to_acc1, "0.5"}, {to_acc2, "0.5"}], "15.00")
{:ok,
%FinancialSystem.Account{
  balance: %FinancialSystem.Money{
    currency: %FinancialSystem.Currency{
      alphabetic_code: "BRL",
      decimal_places: 2,
      numeric_code: 100
    },
    minor_units: 2742,
    precision: 2
  },
  number: 123,
  owner: "Arthur Dent"
},
[
  %FinancialSystem.Account{
    balance: %FinancialSystem.Money{
      currency: %FinancialSystem.Currency{
        alphabetic_code: "BRL",
        decimal_places: 2,
        numeric_code: 100
      },
      minor_units: 17920,
      precision: 3
    },
    number: 423,
    owner: "Ford Prefect"
  },
  %FinancialSystem.Account{
    balance: %FinancialSystem.Money{
      currency: %FinancialSystem.Currency{
        alphabetic_code: "BRL",
        decimal_places: 2,
        numeric_code: 100
      },
      minor_units: 22920,
      precision: 3
    },
    number: 423,
    owner: "Marvin"
  }
]}

withdraw(account, amount)

Withdraws the given amount from the given account if the account has enough funds, i.e., a balance greater than the amount.

Returns a new account with the new balance.

Examples

iex> {:ok, currency} = FinancialSystem.Currency.new("BRL", 100, 2)
iex> {:ok, money} = FinancialSystem.Money.new("42.42", currency)
iex> {:ok, acc} = FinancialSystem.Account.new(123, money, "Arthur Dent")
iex> FinancialSystem.withdraw(acc, "20.00")
{:ok,
%FinancialSystem.Account{
  balance: %FinancialSystem.Money{
    currency: %FinancialSystem.Currency{
      alphabetic_code: "BRL",
      decimal_places: 2,
      numeric_code: 100
    },
    minor_units: 2242,
    precision: 2
  },
  number: 123,
  owner: "Arthur Dent"
}}